Convert more unit tests to mocha/chai/Typescript.
