+++ package.json
@@ -11,9 +11,9 @@
   "engine": "node >= 6.9.1",
   "license": "MIT",
   "scripts": {
-    "test": "nyc --reporter=html mocha --require ts-node/register test/**/*.ts",
+    "test": "nyc --reporter=html mocha --timeout 3000 --require ts-node/register test/**/*.ts",
     "coveralls": "tsc && jscoverage lib && NODETELNETCLIENT_COV=1 nodeunit --reporter=lcov test | coveralls",
-    "lint": "eslint 'src/**/*.ts' 'test/**/*.js'"
+    "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'"
   },
   "dependencies": {
     "net": "^1.0.2",
