+++ test/002_busybox.ts
@@ -1,15 +1,12 @@
-/* eslint-disable dot-notation */
-const { Telnet } = process.env.NODETELNETCLIENT_COV
-  ? require('../lib-cov/index')
-  : require('../lib/index')
-const nodeunit = require('nodeunit')
-const telnet_server = require('telnet')
+import { expect } from 'chai'
+import { Telnet } from '../src'
+import { createServer, Server } from 'net'
 
-let srv
+let server: Server
 
-exports['busybox'] = nodeunit.testCase({
-  setUp: function (callback) {
-    srv = telnet_server.createServer(function (c) {
+describe('busybox', () => {
+  before((done) => {
+    server = createServer(function (c) {
       c.write(Buffer.from('BusyBox v1.19.2 () built-in shell (ash)\n'
         + "Enter 'help' for a list of built-in commands.\n\n/ # ", 'ascii'))
 
@@ -20,18 +17,12 @@ exports['busybox'] = nodeunit.testCase({
       })
     })
 
-    srv.listen(2323, function () {
-      callback()
-    })
-  },
+    server.listen(2323, done)
+  })
 
-  tearDown: function (callback) {
-    srv.close(function () {
-      callback()
-    })
-  },
+  after((done) => server.close(done))
 
-  exec_string_shell_prompt: function (test) {
+  it('exec_string_shell_prompt', (done) => {
     const connection = new Telnet()
     const params = {
       host: '127.0.0.1',
@@ -44,15 +35,15 @@ exports['busybox'] = nodeunit.testCase({
       connection.exec('uptime', function (_err, resp) {
         connection.end().finally()
 
-        test.strictEqual(resp, '23:14  up 1 day, 21:50, 6 users, load averages: 1.41 1.43 1.41\n')
-        test.done()
+        expect(resp).to.equal('23:14  up 1 day, 21:50, 6 users, load averages: 1.41 1.43 1.41\n')
+        done()
       }).finally()
     })
 
     connection.connect(params).finally()
-  },
+  })
 
-  exec_regex_shell_prompt: function (test) {
+  it('exec_regex_shell_prompt', (done) => {
     const connection = new Telnet()
     const params = {
       host: '127.0.0.1',
@@ -65,11 +56,11 @@ exports['busybox'] = nodeunit.testCase({
       connection.exec('uptime', function (_err, resp) {
         connection.end().finally()
 
-        test.strictEqual(resp, '23:14  up 1 day, 21:50, 6 users, load averages: 1.41 1.43 1.41\n')
-        test.done()
+        expect(resp).to.equal('23:14  up 1 day, 21:50, 6 users, load averages: 1.41 1.43 1.41\n')
+        done()
       }).finally()
     })
 
     connection.connect(params).finally()
-  }
+  })
 })
