+++ test/004_negotiation_optional.ts
@@ -1,32 +1,23 @@
-/* eslint-disable dot-notation */
-const { Telnet } = process.env.NODETELNETCLIENT_COV
-  ? require('../lib-cov/index')
-  : require('../lib/index')
-const nodeunit = require('nodeunit')
-const telnet_server = require('telnet')
+import { expect } from 'chai'
+import { Telnet } from '../src'
+import { createServer, Server } from 'net'
 
-let srv
+let server: Server
 
-exports['negotiation_optional'] = nodeunit.testCase({
-  setUp: function (callback) {
-    srv = telnet_server.createServer(function (c) {
+describe('negotiation_optional', () => {
+  before((done) => {
+    server = createServer(function (c) {
       c.on('data', function () {
         c.write(Buffer.from('Hello, user.\n'))
       })
     })
 
-    srv.listen(2323, function () {
-      callback()
-    })
-  },
+    server.listen(2323, done)
+  })
 
-  tearDown: function (callback) {
-    srv.close(function () {
-      callback()
-    })
-  },
+  after((done) => server.close(done))
 
-  send_data: function (test) {
+  it('send_data', (done) => {
     const connection = new Telnet()
     const params = {
       host: '127.0.0.1',
@@ -37,18 +28,18 @@ exports['negotiation_optional'] = nodeunit.testCase({
     connection.on('connect', function () {
       connection.send('Hello, server.', {
         ors: '\r\n',
-        waitfor: '\n'
+        waitFor: '\n'
       }, function (_error, data) {
-        test.strictEqual(data?.toString(), 'Hello, user.\r\n')
-        test.done()
+        expect(data?.toString()).to.equal('Hello, user.\n')
         connection.end().finally()
+        done()
       }).finally()
     })
 
     connection.connect(params).finally()
-  },
+  })
 
-  send_data_without_options: function (test) {
+  it('send_data_without_options', (done) => {
     const connection = new Telnet()
     const params = {
       host: '127.0.0.1',
@@ -59,16 +50,16 @@ exports['negotiation_optional'] = nodeunit.testCase({
 
     connection.on('connect', function () {
       connection.send('Hello, server.', function (_error, data) {
-        test.strictEqual(data.toString(), 'Hello, user.\r\n')
-        test.done()
+        expect(data?.toString()).to.equal('Hello, user.\n')
         connection.end().finally()
+        done()
       }).finally()
     })
 
     connection.connect(params).finally()
-  },
+  })
 
-  send_data_without_options_promise: function (test) {
+  it('send_data_without_options_promise', (done) => {
     const connection = new Telnet()
     const params = {
       host: '127.0.0.1',
@@ -79,10 +70,10 @@ exports['negotiation_optional'] = nodeunit.testCase({
 
     connection.connect(params).then(function () {
       connection.send('Hello, server.').then(function (data) {
-        test.strictEqual(data.toString(), 'Hello, user.\r\n')
-        test.done()
+        expect(data?.toString()).to.equal('Hello, user.\n')
         connection.end().finally()
+        done()
       })
     })
-  }
+  })
 })
